# cmd/docs/swagger.yaml
definitions:
  dto.AuthenticationTokens:
    properties:
      access_token:
        type: string
      access_token_expiry_time:
        type: string
      refresh_token:
        type: string
      refresh_token_expiry_time:
        type: string
    type: object
  dto.CustomerResult:
    properties:
      country:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  dto.LoginParam:
    properties:
      email:
        example: john.doe@example.com
        maxLength: 255
        type: string
    required:
    - email
    type: object
  dto.LoginResult:
    properties:
      authentication_tokens:
        $ref: '#/definitions/dto.AuthenticationTokens'
      customer:
        $ref: '#/definitions/dto.CustomerResult'
    type: object
  dto.ReportResult:
    properties:
      data: {}
      date:
        type: string
      generated_at:
        type: string
      report_type:
        type: string
    type: object
  response.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  response.Problem:
    properties:
      code:
        type: string
      detail:
        type: string
      errors:
        items:
          $ref: '#/definitions/response.FieldError'
        type: array
      instance:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
info:
  contact: {}
  description: 'This is a documentation for Screening Test: Data Automation & Retrieval
    Engineer (PostgreSQL / Go)'
  title: 'Screening Test: Data Automation & Retrieval Engineer (PostgreSQL / Go)'
  version: "1.0"
paths:
  /api/health-check:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Check the API health
      tags:
      - Health
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Customer login using a registered email. Returns an access token
        (valid 1 day) and a refresh token (valid 30 days).
      parameters:
      - description: Login body request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResult'
        "401":
          description: Email is not registered
          schema:
            $ref: '#/definitions/response.Problem'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/response.Problem'
        "429":
          description: Too many requests
          schema:
            $ref: '#/definitions/response.Problem'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Problem'
      summary: Login
      tags:
      - Authentication
  /api/v1/reports:
    get:
      description: 'Get reports based on report type in query (default: report_sales)'
      parameters:
      - default: report_sales
        description: Report type to get
        in: query
        name: report_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReportResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Problem'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Problem'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Problem'
        "404":
          description: Data not found
          schema:
            $ref: '#/definitions/response.Problem'
        "429":
          description: Too many requests
          schema:
            $ref: '#/definitions/response.Problem'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Problem'
      security:
      - BearerAuth: []
      summary: Get reports
      tags:
      - Reports
securityDefinitions:
  BearerAuth:
    description: User API token issued by POST /api/v1/auth/login. Send it as "Bearer
      <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"